se repara migraciones reset

+++ database/migrations/2016_07_06_180221_add_token.php
@@ -27,9 +27,10 @@ class AddToken extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            //
-            $table->dropColumn('token');
-        });
+        if (Schema::hasColumn('users', 'token')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('token');
+            });
+        }
     }
 }
